Change device widget update strategy
Instead of trying to track changes to device attachment state, refresh
the entire state on changes. This takes longer, but I think it's still
acceptable, and seems more reliable. An async queue is used to
debounce attachment events.

Also, redraw the menu if it's being shown as changes happen. It can
get separated from the tray bar right now if devices are removed, or
expand off the screen if they're added, so I still need to figure out
how to reposition it while it's visible.