Revert device state poll
This just adds a check of the VM's attached devices in
device_list_update, so those are properly styled in the menu.

The wait_for_condition function is kept to support moving devices
between VMs.

There seems to be more jitter in redisplaying the menu with this
approach, and the queue did not seem to help, so that was removed for
now.