Update updater tests
Mockups in core-admin-client got fixed default-dvm (it's properly AppVM
now), and also new test-disp (a DispVM). This means there is one more
AppVM, and the total number of VMs also increased. Adjust tests for this
situation.

This change is related to 9c32984 "Late GUID for preloaded disposables"
commit in core-admin-client.